+++ packages/visual-grid-client/test/e2e/takeScreenshot.e2e.test.js
@@ -34,7 +34,7 @@ describe('takeScreenshot e2e', () => {
     const processPageAndSerialize = `(${await getProcessPageAndSerialize()})()`
     await page.goto(website)
     const {cdt, url, resourceUrls, blobs, frames} = await page.evaluate(processPageAndSerialize)
-    const [err, [{imageLocation, renderId}]] = await presult(
+    const [err, result] = await presult(
       takeScreenshot({
         apiKey,
         showLogs: process.env.APPLITOOLS_SHOW_LOGS,
@@ -50,6 +50,8 @@ describe('takeScreenshot e2e', () => {
 
     err && console.log(err)
     expect(err).to.be.undefined
+
+    const [{imageLocation, renderId}] = result
     expect(renderId).to.not.be.undefined
 
     console.log(imageLocation)