+++ packages/dom-snapshot/CHANGELOG.md
@@ -4,6 +4,10 @@
 ## Unreleased
 
 
+## 4.0.0 - 2020/7/26
+
+- support non-200 resources ([Trello](https://trello.com/c/J5lBWutP))
+
 ## 3.7.3 - 2020/7/23
 
 - support src attribute in video tags