+++ packages/eyes-protractor/CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## Unreleased
 
+
+## 1.5.1 - 2020/7/26
+
+- updated to @applitools/eyes-sdk-core@11.5.0 (from 11.4.1)
+- updated to @applitools/visual-grid-client@14.6.0 (from 14.5.15)
+
+## 1.5.0 - 2020/7/24
+
 - remove type definitions
+- updated to @applitools/eyes-sdk-core@11.4.1 (from 11.3.4)
+- updated to @applitools/visual-grid-client@14.5.15 (from 14.5.7)
 
 ## 1.4.0 - 2020/7/8