+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 
 
+## 11.5.0 - 2020/7/26
+
+- support non-200 resources ([Trello](https://trello.com/c/J5lBWutP))
+
+## 11.4.1 - 2020/7/24
+
+- fix issue with duplicate copies of the SDK ([Trello](https://trello.com/c/4Gd7uWtS))
+
 ## 11.4.0 - 2020/7/24
 
 - updated enum types to be more explicit about values