+++ packages/eyes-webdriverio-4-service/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.4.3 - 2020/7/24
+
+- updated to @applitools/eyes.webdriverio@2.20.0 (from 2.18.0)
+
 ## 1.4.2 - 2020/6/29
 
 - updated to @applitools/eyes.webdriverio@2.18.0 (from 2.17.1)